refactor(helpers): update errorFormatter import paths for backward compatibility
- Changed import path for `formatError` in helpers.js and helpers.unit.test.js
- Maintained backward compatibility by adjusting the import to the new location
